+++ octasine/src/audio/parameters/mod.rs
@@ -17,6 +17,7 @@ mod operator_volume;
 use array_init::array_init;
 
 use crate::common::{SampleRate, NUM_LFOS, NUM_OPERATORS};
+use crate::parameters::master_pitch_bend_range::MasterPitchBendRangeValue;
 use crate::parameters::*;
 
 use self::common::{AudioParameter, InterpolatableAudioParameter, SimpleAudioParameter};
@@ -56,6 +57,8 @@ impl<P: AudioParameter> AudioParameterPatchInteraction for P {
 pub struct AudioParameters {
     pub master_volume: MasterVolumeAudioParameter,
     pub master_frequency: MasterFrequencyAudioParameter,
+    pub master_pitch_bend_range_up: SimpleAudioParameter<MasterPitchBendRangeValue>,
+    pub master_pitch_bend_range_down: SimpleAudioParameter<MasterPitchBendRangeValue>,
     pub operators: [OperatorAudioParameters; NUM_OPERATORS],
     pub lfos: [LfoAudioParameters; NUM_LFOS],
 }
@@ -65,6 +68,8 @@ impl Default for AudioParameters {
         Self {
             master_volume: Default::default(),
             master_frequency: Default::default(),
+            master_pitch_bend_range_up: Default::default(),
+            master_pitch_bend_range_down: Default::default(),
             operators: array_init(OperatorAudioParameters::new),
             lfos: array_init(LfoAudioParameters::new),
         }
@@ -79,6 +84,12 @@ macro_rules! impl_patch_interaction {
                 Parameter::Master(p) => match p {
                     MasterParameter::Volume => $f(&mut self.master_volume, input),
                     MasterParameter::Frequency => $f(&mut self.master_frequency, input),
+                    MasterParameter::PitchBendRangeUp => {
+                        $f(&mut self.master_pitch_bend_range_up, input)
+                    }
+                    MasterParameter::PitchBendRangeDown => {
+                        $f(&mut self.master_pitch_bend_range_down, input)
+                    }
                 },
                 Parameter::Operator(index, p) => {
                     use OperatorParameter::*;