+++ octasine/src/gui/mod.rs
@@ -32,6 +32,7 @@ use iced_baseview::{
 use serde::{Deserialize, Serialize};
 
 use crate::common::NUM_OPERATORS;
+use crate::parameters::master_pitch_bend_range::MasterPitchBendRangeValue;
 use crate::parameters::*;
 use crate::sync::GuiSyncHandle;
 
@@ -40,7 +41,9 @@ use operator::OperatorWidgets;
 use patch_picker::PatchPicker;
 use style::Theme;
 
+use self::common::{container_l2, container_l3, space_l3};
 use self::corner::CornerWidgets;
+use self::knob::{master_pitch_bend_range_down, master_pitch_bend_range_up, OctaSineKnob};
 use self::operator::ModTargetPicker;
 use self::style::container::ContainerStyle;
 
@@ -123,6 +126,7 @@ pub enum Message {
     RenamePatch,
     ClearPatch,
     ClearBank,
+    ShowPatchSettings,
     SaveBankOrPatchToFile(PathBuf, Vec<u8>),
     LoadBankOrPatchesFromPaths(Vec<PathBuf>),
     ChangeParameterByTextInput {
@@ -146,6 +150,7 @@ pub enum ModalAction {
         options: Vec<CompactString>,
         choice: CompactString,
     },
+    PatchSettings,
 }
 
 pub struct OctaSineIcedApplication<H: GuiSyncHandle> {
@@ -161,6 +166,8 @@ pub struct OctaSineIcedApplication<H: GuiSyncHandle> {
     lfo_4: LfoWidgets,
     corner: CornerWidgets,
     modal_action: Option<ModalAction>,
+    master_pitch_bend_up: OctaSineKnob<MasterPitchBendRangeValue>,
+    master_pitch_bend_down: OctaSineKnob<MasterPitchBendRangeValue>,
 }
 
 impl<H: GuiSyncHandle> OctaSineIcedApplication<H> {
@@ -170,6 +177,12 @@ impl<H: GuiSyncHandle> OctaSineIcedApplication<H> {
             Parameter::Master(MasterParameter::Volume) => self.corner.master_volume.set_value(v),
             Parameter::Master(MasterParameter::Frequency) => {
                 self.corner.master_frequency.set_value(v)
+            }
+            Parameter::Master(MasterParameter::PitchBendRangeUp) => {
+                self.master_pitch_bend_up.set_value(v)
+            }
+            Parameter::Master(MasterParameter::PitchBendRangeDown) => {
+                self.master_pitch_bend_down.set_value(v)
             }
             outer_p @ Parameter::Operator(index, p) => {
                 self.operator_1.wave_display.set_value(outer_p, v);
@@ -453,6 +466,9 @@ impl<H: GuiSyncHandle> Application for OctaSineIcedApplication<H> {
 
         let corner = CornerWidgets::new(&sync_handle);
 
+        let master_pitch_bend_up = master_pitch_bend_range_up(&sync_handle);
+        let master_pitch_bend_down = master_pitch_bend_range_down(&sync_handle);
+
         let app = Self {
             sync_handle,
             theme: style,
@@ -466,6 +482,8 @@ impl<H: GuiSyncHandle> Application for OctaSineIcedApplication<H> {
             lfo_4,
             corner,
             modal_action: None,
+            master_pitch_bend_up,
+            master_pitch_bend_down,
         };
 
         (app, Command::none())
@@ -777,6 +795,9 @@ impl<H: GuiSyncHandle> Application for OctaSineIcedApplication<H> {
             Message::ClearBank => {
                 self.modal_action = Some(ModalAction::ClearBank);
             }
+            Message::ShowPatchSettings => {
+                self.modal_action = Some(ModalAction::PatchSettings);
+            }
             Message::SaveBankOrPatchToFile(path_buf, bytes) => {
                 if let Err(err) = save_data_to_file(path_buf, bytes) {
                     ::log::error!("Error saving patch/patch bank to file: {:#}", err)
@@ -833,6 +854,7 @@ impl<H: GuiSyncHandle> Application for OctaSineIcedApplication<H> {
                         self.set_value(parameter.parameter(), value_patch, true);
                     }
                 }
+                Some(ModalAction::PatchSettings) => (),
                 None => (),
             },
             Message::ModalSetParameterByChoicesUpdate(new_choice) => {
@@ -894,59 +916,109 @@ impl<H: GuiSyncHandle> Application for OctaSineIcedApplication<H> {
                 ModalAction::SetParameterByChoices { parameter, .. } => {
                     format!("SET {}", parameter.parameter().name().to_uppercase())
                 }
+                ModalAction::PatchSettings => "PATCH SETTINGS".into(),
             };
 
-            let body: Element<'_, Self::Message, Self::Theme> = match modal_action {
-                ModalAction::ClearBank | ModalAction::ClearPatch => Row::new()
-                    .spacing(LINE_HEIGHT / 2)
-                    .width(Length::Fill)
-                    .push(
-                        Button::new(Text::new("YES").horizontal_alignment(Horizontal::Center))
-                            .width(Length::Fill)
-                            .on_press(Message::ModalYes),
-                    )
-                    .push(
-                        Button::new(Text::new("NO").horizontal_alignment(Horizontal::Center))
-                            .width(Length::Fill)
-                            .on_press(Message::ModalClose),
-                    )
-                    .into(),
-                ModalAction::SetParameterByChoices {
-                    options, choice, ..
-                } => Column::new()
-                    .spacing(LINE_HEIGHT)
-                    .push(
-                        PickList::new(options, Some(choice.clone()), |choice| {
-                            Message::ModalSetParameterByChoicesUpdate(choice)
-                        })
-                        .width(Length::Fill),
-                    )
-                    .push(
-                        Row::new()
-                            .spacing(LINE_HEIGHT / 2)
-                            .width(Length::Fill)
-                            .push(
-                                Button::new(
-                                    Text::new("OK").horizontal_alignment(Horizontal::Center),
-                                )
+            match modal_action {
+                ModalAction::ClearBank | ModalAction::ClearPatch => {
+                    let body = Row::new()
+                        .spacing(LINE_HEIGHT / 2)
+                        .width(Length::Fill)
+                        .push(
+                            Button::new(Text::new("YES").horizontal_alignment(Horizontal::Center))
                                 .width(Length::Fill)
                                 .on_press(Message::ModalYes),
-                            )
-                            .push(
-                                Button::new(
-                                    Text::new("CANCEL").horizontal_alignment(Horizontal::Center),
-                                )
+                        )
+                        .push(
+                            Button::new(Text::new("NO").horizontal_alignment(Horizontal::Center))
                                 .width(Length::Fill)
                                 .on_press(Message::ModalClose),
-                            ),
-                    )
-                    .into(),
-            };
+                        );
 
-            Card::new(Text::new(heading), body)
-                .max_width(LINE_HEIGHT as f32 * 16.0)
-                .padding(LINE_HEIGHT as f32)
-                .into()
+                    Card::new(Text::new(heading), body)
+                        .max_width(LINE_HEIGHT as f32 * 16.0)
+                        .padding(LINE_HEIGHT as f32)
+                        .into()
+                }
+                ModalAction::SetParameterByChoices {
+                    options, choice, ..
+                } => {
+                    let body = Column::new()
+                        .spacing(LINE_HEIGHT)
+                        .push(
+                            PickList::new(options, Some(choice.clone()), |choice| {
+                                Message::ModalSetParameterByChoicesUpdate(choice)
+                            })
+                            .width(Length::Fill),
+                        )
+                        .push(
+                            Row::new()
+                                .spacing(LINE_HEIGHT / 2)
+                                .width(Length::Fill)
+                                .push(
+                                    Button::new(
+                                        Text::new("OK").horizontal_alignment(Horizontal::Center),
+                                    )
+                                    .width(Length::Fill)
+                                    .on_press(Message::ModalYes),
+                                )
+                                .push(
+                                    Button::new(
+                                        Text::new("CANCEL")
+                                            .horizontal_alignment(Horizontal::Center),
+                                    )
+                                    .width(Length::Fill)
+                                    .on_press(Message::ModalClose),
+                                ),
+                        );
+
+                    Card::new(Text::new(heading), body)
+                        .max_width(LINE_HEIGHT as f32 * 16.0)
+                        .padding(LINE_HEIGHT as f32)
+                        .into()
+                }
+                ModalAction::PatchSettings => {
+                    let body = Column::new()
+                        .push(container_l2(
+                            Column::new()
+                                .push(Space::with_height(LINE_HEIGHT))
+                                .push(
+                                    Text::new("GLOBAL PITCH BEND RANGE")
+                                        .size(FONT_SIZE + FONT_SIZE / 2)
+                                        .font(self.theme.font_heading())
+                                        .horizontal_alignment(Horizontal::Center)
+                                        .width(Length::Fixed(f32::from(LINE_HEIGHT * 8))),
+                                )
+                                .push(
+                                    Row::new()
+                                        .push(container_l3(
+                                            self.master_pitch_bend_up.view(&self.theme),
+                                        ))
+                                        .push(space_l3())
+                                        .push(container_l3(
+                                            self.master_pitch_bend_down.view(&self.theme),
+                                        )),
+                                ),
+                        ))
+                        .push(Space::with_height(LINE_HEIGHT))
+                        .push(
+                            Row::new()
+                                .spacing(LINE_HEIGHT / 2)
+                                .width(Length::Fill)
+                                .push(
+                                    Button::new(
+                                        Text::new("CLOSE").horizontal_alignment(Horizontal::Center),
+                                    )
+                                    .width(Length::Fill)
+                                    .on_press(Message::ModalYes),
+                                ),
+                        );
+                    Card::new(Text::new(heading), body)
+                        .max_width(LINE_HEIGHT as f32 * 12.0)
+                        .padding(LINE_HEIGHT as f32)
+                        .into()
+                }
+            }
         })
         .backdrop(Message::ModalClose)
         .on_esc(Message::ModalClose)