+++ octasine/src/parameters/mod.rs
@@ -9,6 +9,7 @@ pub mod lfo_shape;
 pub mod lfo_target;
 pub mod list;
 pub mod master_frequency;
+pub mod master_pitch_bend_range;
 pub mod master_volume;
 pub mod operator_active;
 pub mod operator_envelope;
@@ -96,6 +97,8 @@ impl Parameter {
             Self::None => "None".into(),
             Self::Master(MasterParameter::Frequency) => "Master frequency".into(),
             Self::Master(MasterParameter::Volume) => "Master volume".into(),
+            Self::Master(MasterParameter::PitchBendRangeUp) => "Pitch bend range (up)".into(),
+            Self::Master(MasterParameter::PitchBendRangeDown) => "Pitch bend range (down)".into(),
             Self::Operator(index, p) => match p {
                 OperatorParameter::Volume => format_compact!("OP {} vol", index + 1),
                 OperatorParameter::Active => format_compact!("OP {} active", index + 1),
@@ -156,6 +159,12 @@ impl Parameter {
             Self::None => "None".into(),
             Self::Master(MasterParameter::Frequency) => "Master frequency".into(),
             Self::Master(MasterParameter::Volume) => "Master volume".into(),
+            Self::Master(MasterParameter::PitchBendRangeUp) => {
+                "Master pitch bend range (up)".into()
+            }
+            Self::Master(MasterParameter::PitchBendRangeDown) => {
+                "Master pitch bend range (down)".into()
+            }
             Self::Operator(index, p) => match p {
                 OperatorParameter::Volume => format!("OP {} vol", index + 1),
                 OperatorParameter::Active => format!("OP {} active", index + 1),